started to replace float with double removed previously added shader and replaced it with in-CPU calculations, so we can use doubles (and are not dependent on modern hardware)

+++ Project/core/src/main/java/dhbw/karlsruhe/it/solar/core/ai/AIOutput.java
@@ -1,19 +1,20 @@
 package dhbw.karlsruhe.it.solar.core.ai;
 
-import com.badlogic.gdx.math.Vector2;
+
+import mikera.vectorz.Vector2;
 
 /**
  * Created by Arga on 13.02.2015.
  */
 public class AIOutput {
     private Vector2 position;
-    private float rotation;
+    private double rotation;
 
     public void setPosition(Vector2 newPosition) {
         position = newPosition;
     }
 
-    public void setRotation(float newRotation) {
+    public void setRotation(double newRotation) {
         rotation = newRotation;
     }
 
@@ -21,7 +22,7 @@ public class AIOutput {
         return position;
     }
 
-    public float getRotation() {
+    public double getRotation() {
         return rotation;
     }
 }
